+++ src/index.ts
@@ -1,4 +1,5 @@
 export * from "./model/cellRef";
+export * from "./model/CellFormat";
 export * from "./model/SheetClipboard";
 export * from "./model/SheetHistory";
 export * from "./model/SheetModel";